+++ phpmyfaq/translations/language_de.php
@@ -1290,13 +1290,13 @@ $PMF_LANG['buttonCheckHealth'] = 'Systemzustand jetzt prüfen';
 $PMF_LANG['buttonCheckUpdates'] = 'Suche jetzt nach Updates';
 $PMF_LANG['buttonDownloadPackage'] = 'Jetzt herunterladen';
 $PMF_LANG['buttonExtractPackage'] = 'Paket jetzt entpacken';
-$PMF_LANG['versionIsUpToDate'] = '✅ Deine installierte Version ist aktuell.';
-$PMF_LANG['healthCheckOkay'] = '✅ Deine installierte Version ist in Ordnung.';
-$PMF_LANG['downloadSuccessful'] = '✅ Download-Paket erfolgreich heruntergeladen!';
-$PMF_LANG['extractSuccessful'] = '✅ Paket erfolgreich entpackt!';
-$PMF_LANG['downloadFailure'] = '❌ Download-Paket konnte nicht heruntergeladen werden';
-$PMF_LANG['verificationFailure'] = '❌ Download-Paket konnte nicht verifiziert werden';
-$PMF_LANG['extractFailure'] = '❌ Das heruntergeladene Paket konnte nicht entpackt werden.';
+$PMF_LANG['versionIsUpToDate'] = 'Deine installierte Version ist aktuell.';
+$PMF_LANG['healthCheckOkay'] = 'Deine installierte Version ist in Ordnung.';
+$PMF_LANG['downloadSuccessful'] = 'Download-Paket erfolgreich heruntergeladen!';
+$PMF_LANG['extractSuccessful'] = 'Paket erfolgreich entpackt!';
+$PMF_LANG['downloadFailure'] = 'Download-Paket konnte nicht heruntergeladen werden';
+$PMF_LANG['verificationFailure'] = 'Download-Paket konnte nicht verifiziert werden';
+$PMF_LANG['extractFailure'] = 'Das heruntergeladene Paket konnte nicht entpackt werden.';
 $PMF_LANG['msgCurrentVersion'] = 'Aktuelle Version: ';
 $PMF_LANG['msgLastCheckDate'] = 'Letzte Überprüfung auf Aktualisierungen: ';
 $PMF_LANG['msgLastVersionAvailable'] = 'Aktuellste verfügbare Version: ';
@@ -1377,7 +1377,7 @@ $PMF_LANG['msgUserList'] = 'Benutzerliste';
 
 // added v4.0.0-alpha.2 - 2024-04-30 by Thorsten
 $PMF_LANG['msgNoQuestionAndAnswer'] = 'Keine Frage und Antwort gefunden.';
-$PMF_LANG['msgNotInMaintenanceMode'] = '⚠️Die FAQ ist nicht im Wartungs-Modus.';
+$PMF_LANG['msgNotInMaintenanceMode'] = '⚠️ Die FAQ ist nicht im Wartungs-Modus.';
 
 // added v4.0.0-alpha.3 - 2024-06-03 by Jan
 $PMF_LANG['msgSeparateKeywordsWithCommas'] = 'Mehrere Suchbegriffe müssen mit Kommas getrennt werden.';